tests/wpt/web-platform-tests/fetch/metadata/generated/worker-dedicated-constructor.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/worker-dedicated-constructor.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for dedicated worker via the "Worker" constructor</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <script type="module">
  'use strict';
  function induceRequest(url, options) {
    return new Promise((resolve, reject) => {
      const worker = new Worker(url, options);
      worker.onmessage = resolve;
      worker.onerror = reject;
    });
  }

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['same-origin']);
        });
  }, 'sec-fetch-mode - no options');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url, {"type": "module"})
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['same-origin']);
        });
  }, 'sec-fetch-mode - options: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_array_equals(headers['sec-fetch-dest'], ['worker']);
        });
  }, 'sec-fetch-dest - no options');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url, {"type": "module"})
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_array_equals(headers['sec-fetch-dest'], ['worker']);
        });
  }, 'sec-fetch-dest - options: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - no options');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url, {"type": "module"})
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - options: type=module');
  </script>
</html>
